server/requirements.txt
Flask==1.1.1
Flask-Testing==0.7.1
Flask-SQLAlchemy==2.4.0
psycopg2==2.8.3
psycopg2-binary==2.8.3
flask-cors==3.0.8
PyJWT